Loading Infrastructure
Fixed loading files to ImageData, ImageFields and MaskFields. TODO: create and connect fitter
