+++ src/ImageFields.py
@@ -1,6 +1,9 @@
+import typing
+from PyQt5 import QtCore
 from PyQt5.QtWidgets import *
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QPalette, QColor, QIcon
+from PyQt5.QtWidgets import QWidget
 from pyqtgraph import PlotWidget, plot
 import pyqtgraph as pg
 import sys
@@ -20,6 +23,60 @@ import os
 import scipy
 import ImageData
 import Mainapp
+
+class ImFields(QMainWindow):
+    def __init__(self):
+        super(ImFields, self).__init__()
+        self.central_widget = QWidget()
+        self.layoutV1 = QVBoxLayout()
+        self.layoutH1 = QHBoxLayout() # Load/Save Image
+        self.layoutH2 = QHBoxLayout() # peak nums
+        self.layoutH3 = QHBoxLayout() # max/min pixels
+        self.layoutH3b = QHBoxLayout() # max/min pixels
+        self.layoutH4 = QHBoxLayout() # Final Prompts
+        self.setCentralWidget(self.central_widget)
+
+        loadImagePrompt = QPushButton('Load Image')
+        saveImagePrompt = QPushButton('*Save Image*')
+        loadImagePrompt.clicked.connect(ImageData.ImageReader.on_LoadIm_clicked)
+        saveImagePrompt.clicked.connect(ImageData.ImageReader.on_SaveIm_clicked)
+
+        ClearPrompt = QPushButton('*Toggle Image Additives*') #TODO
+        ReducePrompt = QPushButton('*Reduce Peak Number*')#TODO
+
+        ImFields.xpeaksIn = x_peak_read()
+        ImFields.ypeaksIn = y_peak_read()
+        ImFields.xminIn = x_min_read()
+        ImFields.yminIn = y_min_read()
+        ImFields.xmaxIn = x_max_read()
+        ImFields.ymaxIn = y_max_read()
+
+        self.central_widget.setLayout(self.layoutV1)
+        self.layoutV1.addLayout(self.layoutH1)
+        self.layoutH1.addWidget(loadImagePrompt)
+        self.layoutH1.addWidget(saveImagePrompt)
+
+        self.layoutV1.addLayout(self.layoutH2)
+        self.layoutH2.addWidget(QLabel('X-peaks'))
+        self.layoutH2.addWidget(ImFields.xpeaksIn)
+        self.layoutH2.addWidget(QLabel('Y-peaks'))
+        self.layoutH2.addWidget(ImFields.ypeaksIn)
+
+        self.layoutV1.addLayout(self.layoutH3)
+        self.layoutH3.addWidget(QLabel('Min X'))
+        self.layoutH3.addWidget(ImFields.xminIn)
+        self.layoutH3.addWidget(QLabel('Max X'))
+        self.layoutH3.addWidget(ImFields.xmaxIn)
+
+        self.layoutV1.addLayout(self.layoutH3b)
+        self.layoutH3b.addWidget(QLabel('Min Y'))
+        self.layoutH3b.addWidget(ImFields.yminIn)
+        self.layoutH3b.addWidget(QLabel('Max Y'))
+        self.layoutH3b.addWidget(ImFields.ymaxIn)
+        
+        self.layoutV1.addLayout(self.layoutH4)
+        self.layoutH4.addWidget(ClearPrompt)
+        self.layoutH4.addWidget(ReducePrompt)
 
 class y_peak_read(QLineEdit):
     def __init__(self):
@@ -207,91 +264,3 @@ class y_max_read(QLineEdit):
         # print(s)
         return
 
-class hole_diam_read(QLineEdit):
-    def __init__(self):
-        QLineEdit.__init__(self)
-        # self.setMaxLength(6)
-        self.setPlaceholderText("mm")
-        self.returnPressed.connect(self.return_pressed)
-        self.selectionChanged.connect(self.selection_changed)
-        self.textChanged.connect(self.text_changed)
-        self.textEdited.connect(self.text_edited)
-
-    def on_button_clicked():
-        alert = QMessageBox()
-        alert.setText('You clicked the button!')
-        alert.exec()
-    
-    def return_pressed(self):
-        print("Return pressed!")
-        
-    def selection_changed(self):
-        return
-
-    def text_changed(self, s):
-        # print("hole diam changed...")
-        # print(s)
-        return
-    
-    def text_edited(self, s):
-        # print("hole diam edited...")
-        # print(s)
-        return
-
-class hole_sep_read(QLineEdit):
-    def __init__(self):
-        QLineEdit.__init__(self)
-        self.setMaxLength(6)
-        self.setPlaceholderText("mm")
-        self.returnPressed.connect(self.return_pressed)
-        self.selectionChanged.connect(self.selection_changed)
-        self.textChanged.connect(self.text_changed)
-        self.textEdited.connect(self.text_edited)
-
-    def on_button_clicked():
-        alert = QMessageBox()
-        alert.setText('You clicked the button!')
-        alert.exec()
-    
-    def return_pressed(self):
-        print("Return pressed!")
-        
-    def selection_changed(self):
-        return
-
-    def text_changed(self, s):
-        return
-    
-    def text_edited(self, s):
-        return
-
-class hole_num_read(QLineEdit):
-    def __init__(self):
-        QLineEdit.__init__(self)
-        self.setMaxLength(2)
-        self.setPlaceholderText("#")
-        self.returnPressed.connect(self.return_pressed)
-        self.selectionChanged.connect(self.selection_changed)
-        self.textChanged.connect(self.text_changed)
-        self.textEdited.connect(self.text_edited)
-
-    def on_button_clicked():
-        alert = QMessageBox()
-        alert.setText('You clicked the button!')
-        alert.exec()
-    
-    def return_pressed(self):
-        print("Return pressed!")
-        
-    def selection_changed(self):
-        return
-
-    def text_changed(self, s):
-        # print("num changed...")
-        # print(s)
-        return
-    
-    def text_edited(self, s):
-        # print("num edited...")
-        # print(s)
-        return
