+++ src/Sliders.py
@@ -59,12 +59,12 @@ class slider1(QWidget):#Threshold
         s = int(self.slValue.text())
         self.sl.setValue(s)
         if threshbool == True:
-            ImageData.changeThreshold(self.returnValue())
+            ImageData.ImageReader.changeThreshold(self.returnValue())
 
     def changeSlider(self):
         self.slValue.setText(str(self.sl.value()))
         if threshbool == True:
-            ImageData.changeThreshold(self.returnValue())
+            ImageData.ImageReader.changeThreshold(self.returnValue())
 
     def returnValue(self):
         return(int(self.slValue.text()))
@@ -107,12 +107,12 @@ class slider2(QWidget):#Prominence
         s = int(self.slValue.text())
         self.sl.setValue(s)
         if prombool == True:
-            ImageData.changeProminence(self.returnValue())
+            ImageData.ImageReader.changeProminence(self.returnValue())
 
     def changeSlider(self):
         self.slValue.setText(str(self.sl.value()))
         if prombool == True:
-            ImageData.changeProminence(self.returnValue())
+            ImageData.ImageReader.changeProminence(self.returnValue())
 
     def returnValue(self):
         return(int(self.slValue.text()))
